requirements.txt
Flask==2.3.3
openai==1.95.1
python-dotenv==1.0.0
numpy==1.24.3
scikit-learn==1.3.0
psycopg2-binary==2.9.7 